requirements.txt
accelerate
pymongo
python-dotenv
openai
Pillow
pyheif